// kservices_ts_api/src/api/domain/entities/sme.entity.ts
export type recordingGetRequest = {
  id: number;
  isDownload: string;
  initialRecord: number;
  batchSize: number;
  startDate: string;
  startDate_op: string;
  endDate: string;
  endDate_op: string;
  duration: string;
  duration_op: string;
  customerNumber: string;
  customerNumber_op: string;
  agentId: string;
  agentId_op: string;
  agentName: string;
  agentName_op: string;
};